If user_vars.yaml does not exist, automatically create it.

+++ lib/commands/init.rb
@@ -12,12 +12,11 @@ module Homeconfig
 
         if Dir.exist?(dotfile_dir)
           puts "Directory already exists: #{dotfile_dir}"
-          return
+        else
+          # 创建 ~/.mydotfiles 目录
+          FileUtils.mkdir_p(dotfile_dir)
+          puts "Created directory: #{dotfile_dir}"
         end
-
-        # 创建 ~/.mydotfiles 目录
-        FileUtils.mkdir_p(dotfile_dir)
-        puts "Created directory: #{dotfile_dir}"
 
         # 生成初始的 user_vars.yaml 文件
         initial_user_vars = {
@@ -29,15 +28,19 @@ module Homeconfig
           },
         }
 
-        # Write the YAML data to user_vars.yaml
-        File.write(user_vars_file, YAML.dump(initial_user_vars))
-        puts "Created initial user_vars.yaml: #{user_vars_file}"
-
-        # Create the .gitignore file and add user_vars.yaml to it
-        File.open(gitignore_file, "w") do |file|
-          file.puts "user_vars.yaml"
+        if !File.exist?(user_vars_file)
+          # Write the YAML data to user_vars.yaml
+          File.write(user_vars_file, YAML.dump(initial_user_vars))
+          puts "Created initial user_vars.yaml: #{user_vars_file}"
         end
-        puts "Created .gitignore and added user_vars.yaml to it: #{gitignore_file}"
+
+        if !File.exist?(gitignore_file)
+          # Create the .gitignore file and add user_vars.yaml to it
+          File.open(gitignore_file, "w") do |file|
+            file.puts "user_vars.yaml"
+          end
+          puts "Created .gitignore and added user_vars.yaml to it: #{gitignore_file}"
+        end
       end
 
       def self.help
